Compile modules that are needed by template haskell, even with -fno-code.
Summary:
This patch relates to Trac #8025

The goal here is to enable typechecking of packages that contain some
template haskell. Prior to this patch, compilation of a package with -fno-code
would fail if any functions in the package were called from within a splice.

downsweep is changed to do an additional pass over the modules, targetting
any ModSummaries transitively depended on by a module that has
LangExt.TemplateHaskell enabled. Those targeted modules have hscTarget changed
from HscNothing to the default target of the platform.

We use temporary files for .o and .hi outputs, and disable optimizations.

There is a small change to the prevailing_target logic to enable this.

There is a change to DriverPipeline.getLocation. Previously a completely new
ModLocation would be returned, derived from the input filename and the dynflags.
Now, if there is an existing ModLocation, it is modified and returned.

A simple test is added.

A note [-fno-code mode] is added in GhcMake.hs and several comments are updated.

A couple of existing too-long lines are reformatted.

Reviewers: austin, bgamari

Subscribers: rwbarton, thomie

GHC Trac Issues: #8025

Differential Revision: https://phabricator.haskell.org/D3441