Refactor temp files cleanup
Summary:
Remove filesToNotIntermediateClean from DynFlags, create a data type
FilesToClean, and change filesToClean in DynFlags to be a FilesToClean.

Move temporary file functions from SysTools to a new module FileCleanup in
compiler/main. To facilitate this, traceMsg is moved to ErrUtils.

Modify FileCleanup.newTempName and the Temporary constructor of
PipelineMonad.PipelineOutput to take a TempFileLifetime, which specifies
whether a temp file should live until the end of GhcMonad.withSession, or until
the next time cleanShortLivedTempFiles is called. Replace calls of
cleanTempFilesExcept with calls to cleanShortLivedTempFiles.

These changes allow the cleaning of short lived files in GhcMake to be much
more efficient, since the subset of files to delete does not have to be
recalculated on every call to cleanTempFilesExcept.

HscTypes.hptObjs is removed as it is no longer used.

A new performance test T13701 is added, which passes both with and without
-keep-tmp-files. Note that there are still at two hotspots caused by
algorithms quadratic in the number of modules, however neither of them allocate.
They are:
* DriverPipeline.compileOne'.needsLinker
* GhcMake.getModLoop

DriverPipeline.compileOne'.needsLinker is changed slightly to improve the
constant factors, but it is still quadratic in the number of modules in the
worst case.

One type, FilesToClean, and one function emptyFilesToClean is added to
DynFlags.hs, since they are needed to define and initialize a DynFlags.

Reviewers: austin, hvr, bgamari

Subscribers: rwbarton, thomie

GHC Trac Issues: #13701

Differential Revision: https://phabricator.haskell.org/D3620