Change HscEnv.hsc_HPT from HomePackageTable to IORef HomePackageTable
Summary:
This Revision is a work in progress, and I am mainly submitting it at
this point for validation. There will be several more Diffs building
on this one to implement the design described in the ticket.

This should result in no change in behaviour.

This change is to support #14095: Improve parallelism in --make mode

Reviewers: austin, hvr, bgamari

Subscribers: rwbarton, thomie

GHC Trac Issues: #14095

Differential Revision: https://phabricator.haskell.org/D3832